requirements.txt
asgiref==3.7.2
astroid==2.15.5
colorama==0.4.6
dill==0.3.6
Django==4.2.2
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
tzdata==2023.3
wrapt==1.15.0
